+++ chroma-league/src/main/java/com/bonepl/chromaleague/rest/http/client/BlockingResponseInterceptor.java
@@ -1,4 +1,4 @@
-package com.bonepl.chromaleague.rest.http;
+package com.bonepl.chromaleague.rest.http.client;
 
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpResponseInterceptor;
