+++ chroma-league/src/main/java/com/bonepl/chromaleague/rest/http/handlers/GameStatsResponseHandler.java
@@ -1,4 +1,4 @@
-package com.bonepl.chromaleague.rest.http;
+package com.bonepl.chromaleague.rest.http.handlers;
 
 import com.bonepl.chromaleague.rest.gamestats.GameStats;
 import com.jsoniter.JsonIterator;
