+++ chroma-league/src/main/java/com/bonepl/chromaleague/tasks/FetchActivePlayerTask.java
@@ -1,7 +1,7 @@
 package com.bonepl.chromaleague.tasks;
 
-import com.bonepl.chromaleague.rest.http.ActivePlayerResponseHandler;
 import com.bonepl.chromaleague.rest.http.LeagueHttpClients;
+import com.bonepl.chromaleague.rest.http.handlers.ActivePlayerResponseHandler;
 import com.bonepl.chromaleague.state.RunningState;
 
 public class FetchActivePlayerTask implements Runnable {
